salvando nova transaction, validacoes l1,l2,l3 e l4

+++ src/main/kotlin/com/caju/teste/autorizador/adapter/output/request/TransactionRequest.kt
@@ -1,8 +1,10 @@
 package com.caju.teste.autorizador.adapter.output.request
 
+import java.math.BigDecimal
+
 data class TransactionRequest(
         val account : String,
-        val totalAmount: Double,
+        val totalAmount: BigDecimal,
         val mcc: String,
         val merchant: String
 )
