+++ src/main/kotlin/com/caju/teste/autorizador/domain/model/Transaction.kt
@@ -1,8 +1,10 @@
 package com.caju.teste.autorizador.domain.model
 
+import java.math.BigDecimal
+
 class Transaction(
         val accountId: String,
-        val totalAmount: Double,
+        val totalAmount: BigDecimal,
         val mcc: String,
         val merchant: String
 )
